Option in Fit function and header file for FitPanel class
